requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
python-json-logger==0.1.11
PyYAML==5.1.2
regex==2020.5.14
six==1.12.0
structlog==19.2.0
toml==0.10.1
typed-ast==1.4.1
uWSGI==2.0.18
viaa-chassis==0.2.0rc1
wcwidth==0.1.9
Werkzeug==0.16.0
